+++ front-end-tools/CHANGELOG.md
@@ -1,7 +1,20 @@
 ## Unreleased changes
 
+## 2.0.0
+
+- Add support for getting the embedded schema from the module.
+- Display schema template to user.
+
+## 1.3.0
+
 - Add link to source code.
+
+## 1.2.0
+
 - Add `smart-contract-name` dropDown. The `smart-contract-names` are extracted from the module.
+
+## 1.1.0
+
 - Add link to developer documentation
 
 ## 1.0.0